[CL-2107] Move functions to util file
